Move Bitwarden tests to own file
